+++ src/view/util.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING, Union, overload
 
 from ._logging import Internal, Service
 from ._util import shell_hint
-from .exceptions import AppNotFoundError, EnvironmentError, MistakeError
+from .exceptions import AppNotFoundError, BadEnvironmentError, MistakeError
 
 if TYPE_CHECKING:
     from .app import App
@@ -109,7 +109,7 @@ def env(key: str, *, tp: type[EnvConv] = str) -> EnvConv:
     value = os.environ.get(key)
 
     if not value:
-        raise EnvironmentError(
+        raise BadEnvironmentError(
             f'environment variable "{key}" not set',
             hint=shell_hint(
                 f"set {key}=..." if os.name == "nt" else f"export {key}=..."