+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for implicit query parameters
 - Renamed `debug` to `enable_debug`
 - Added `debug`, `info`, `warning`, `error`, and `critical` logging functions
+- Added `InvalidRouteError`, `DuplicateRouteError`, `ViewInternalError`, and `ConfigurationError`
+- Renamed `EnvironmentError` to `BadEnvironmentError`
 
 ## [1.0.0-alpha5] - 2023-09-24